Utility Class Improvement and Standardization
Since all methods are static, and we are not interested in instantiating the class, why not make it a proper utility class by making it have a private constructor and explicitly final. This will make it resemble java.lang.Math. (http://www.docjar.com/html/api/java/lang/Math.java.html)

I also standardized all of the braces because there was type mixing.