Make it unabusable
I'd say its best to create the constants in the actual class where they are most relevant, but if this has to be done, then it should at least be safe from instantiation and from having child classes extending it. If you accept this change, do note that it will need to be a static import in other classes when being used.