Treat Vector3 like struct
This is not a 100% necessary change, but it will improve the easy of usability of these Vectors. Using the builder class, you can say `new Vector.Builder().x(5).y(4).z(5).build()` to build a vector, and then to call the vector's variables, you can directly access them: `vector.x`. This will cause the class to be treated more like a struct would be in C/C++, which would make sense considering that it's only purpose is to hold data. Hence, the purpose of this change is ease of usability. 

NOTE: I was not able to run this on my local machine, so there may be some small errors you need to fix because I didn't get to see any compiler warnings or run-time errors. I recommend checking that out if you are interested in merging this pull.